Update to 3.1.0 (#10)
* Add Compress and NoMapping options to persistence
* Import: Minor code layout cleanup
* Minor updates to comments
* Minor updates to text lines
* Update documentation
* Update tests
* Import: Change data conflict to warning instead of error
* Import: Fix counters resetting if multiple files from pipeline
* Change checking of persistence path to shared function
* Import/Export: Add config object for metadata

+++ src/Disable-XRefPersistence.ps1
@@ -7,7 +7,10 @@ function Disable-XRefPersistence {
         [switch]$Quiet,
 
         [Parameter()]
-        [switch]$Remove
+        [switch]$Remove,
+
+        [Parameter(DontShow)]
+        [switch]$_Reset
     )
 
     Process {
@@ -24,8 +27,8 @@ function Disable-XRefPersistence {
                     Write-Host "No StreamXRef persistence files to delete."
                 }
             }
-            else {
-                # Add ".bak" to name to prevent auto-loading
+            elseif (-not $_Reset -and (Test-Path $PersistPath)) {
+                # Add ".bak" to name to prevent auto-loading (except when resetting)
                 Move-Item $PersistPath "$PersistPath.bak" -Force
             }
 
@@ -44,7 +47,7 @@ function Disable-XRefPersistence {
             $script:PersistEnabled = $false
         }
         else {
-            Write-Error "Unable to determine Application Data path"
+            Write-Error "Data persistence path is not set."
         }
     }
 }
